LayoutTests/fast/page-color-sampling/color-sampling-ignores-images-expected.txt


PASS colors.top is "rgb(255, 99, 71)"
PASS colors.left is null
PASS colors.right is null
PASS colors.bottom is "rgb(211, 211, 211)"
PASS successfullyParsed is true

TEST COMPLETE
Bottom bar
